If output is a future, wait for it to finish.
If a user wants to use a callback and not wait for a program to finish, he can
use :background to run the whole thing in a future. In fact, :background might
go away in the future, seeing as how it is less characters to just wrap the
program call in a future manually.